Refactor: Organize combatant hooks and extract shared test helpers
- Move combatant-related hooks to src/hooks/combatantHooks/ folder
- Extract getTestGlobals and flushAsync to shared tests/helpers.ts
- Rename getShouldBeDefeatedFromActorState to getShouldBeDefeatedFromCombatant

Addresses PR review feedback from @Fronix.
