+++ src/hooks/combatantHooks/combatantDefeatSync.test.ts
@@ -1,4 +1,5 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest';
+import { flushAsync, getTestGlobals } from '../../../tests/helpers.js';
 import {
 	type CombatDefeatSyncTestGlobals,
 	createHasPropertyMock,
@@ -6,15 +7,10 @@ import {
 	createMockCombat,
 	createMockCombatActor,
 	createMockCombatant,
-	getTestGlobals,
-} from '../../tests/mocks/combat.js';
+} from '../../../tests/mocks/combat.js';
 
 function globals() {
 	return getTestGlobals<CombatDefeatSyncTestGlobals>();
-}
-
-async function flushAsync() {
-	await new Promise((resolve) => setTimeout(resolve, 0));
 }
 
 describe('registerCombatantDefeatSync', () => {
